Add logging for ProcessGroup backends.
This PR adds logging to track usage of ProcessGroup backends. The
change involves adding an `init()` method to the ProcessGroup base class and
modifying all the subclasses to call it.

We can't add logging directly in the ProcessGroup constructor since
`getBackendName()` is pure virtual and can't be called from the base class
constructor. See
https://isocpp.org/wiki/faq/strange-inheritance#calling-virtuals-from-ctors for
more details.

Differential Revision: [D34596295](https://our.internmc.facebook.com/intern/diff/D34596295/)

ghstack-source-id: 150381852
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73702