[JIT][WIP] C10_EXPORT -> TORCH_API
These all appear to be defined in libtorch_cpu.so, so they should be marked with TORCH_API. TORCH_API means that these symbols are exported from libtorch_cpu.so and no other libraries. In comparison, C10_EXPORT will export the symbol in _all_ built libraries, if it's available.

I think most of these were fine because most were only defined in cpp files (which would only be included in the targets for one .so file). However, the change in pass_manager.h affects behavior, since the class is defined in the .h file, which could result in two separate implementations of the same static functions. Previously we saw issues on windows with this: https://github.com/pytorch/pytorch/pull/73742

ghstack-source-id: 73592b7fe51893e6bcb0f9b4b9435a09755cfb63
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73818