[c10d] Use c10::Synchronized in c10d::SequenceNum
The new `c10::Synchronized` class helps enforce that locks are properly
held when touching guarded fields. Use it instead of a raw mutex.

Also add an apparently-missing TORCH_CHECK in get(). The header comment
says that it will throw if the number is not set, but get() didn't do
the check that the other methods did.

ghstack-source-id: 0a0fd49d7d6483586ca08925fe2dd39d46f86508
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74049